move runtime memory to separate file
